multithreading in  FLASHProFilter

+++ src/openms/include/OpenMS/ANALYSIS/TOPDOWN/FLASHProFilterAlgorithm.h
@@ -66,7 +66,7 @@ namespace OpenMS
 
     ~FLASHProFilterAlgorithm();
 
-    std::map<int, double> getScores(MSSpectrum &decovSpec, double intThreshold = .0);
+    std::map<double, int> getScores(MSSpectrum &decovSpec, double intThreshold = .0);
 
   protected:
     std::vector<FASTAFile::FASTAEntry> fastaEntry;
